+++ src/examples/http-get-with-query-string.ts
